Added better charge logic / Adapted for slight circuit change
In order for the arduino to read the battery voltage (and therefor percentage) a DPDT switch is required to break the  connection between BAT and A0 when the circuit is switched off. Otherwise the arduino will draw power via A0 when switched off which is not good.

I also switched to an RGB led for the charge indicator LED (only using red and green). If you want to only use a red LED then set `SET_CHARGE_INDICATOR_RGB` to `0` in the code (currently line 40).